+++ apps/cli/src/live-common-setup-base.ts
@@ -41,7 +41,6 @@ setSupportedCurrencies([
   "vertcoin",
   "peercoin",
   "viacoin",
-  "stakenet",
   "bitcoin_testnet",
   "ethereum_ropsten",
   "ethereum_goerli",
@@ -74,7 +73,7 @@ const { format } = winston;
 const { combine, json } = format;
 const winstonFormatJSON = json();
 const winstonFormatConsole = combine(
-  format(({ type: _type, id: _id, date: _date, ...rest }) => rest)(),
+  format(({ type, message, id: _id, date: _date, ...rest }) => ({ ...rest, message: `${type}: ${message}` }) )(),
   format.colorize(),
   simple()
 );