+++ apps/ledger-live-desktop/src/live-common-set-supported-currencies.js
@@ -40,7 +40,6 @@ setSupportedCurrencies([
   "vertcoin",
   "peercoin",
   "viacoin",
-  "stakenet",
   "bitcoin_testnet",
   "ethereum_ropsten",
   "ethereum_goerli",