+++ apps/ledger-live-desktop/src/renderer/components/SyncOnboarding/Manual/SoftwareCheckStep.tsx
@@ -1,15 +1,17 @@
-import React, { useEffect, useState } from "react";
+import React, { useCallback, useEffect, useState } from "react";
 import { Box } from "@ledgerhq/react-ui";
 import { useSelector } from "react-redux";
 
 import { useGenuineCheck } from "@ledgerhq/live-common/hw/hooks/useGenuineCheck";
 import { useGetLatestAvailableFirmware } from "@ledgerhq/live-common/hw/hooks/useGetLatestAvailableFirmware";
+import { DeviceModelId } from "@ledgerhq/types-devices";
 import { command } from "~/renderer/commands";
 import { getCurrentDevice } from "~/renderer/reducers/devices";
 
 import SoftwareCheckContent from "./SoftwareCheckContent";
 import GenuineCheckModal from "./GenuineCheckModal";
-import GenuineCheckAnimationModal from "./GenuineCheckAnimationModal";
+import SoftwareCheckLockedDeviceModal from "./SoftwareCheckLockedDeviceModal";
+import SoftwareCheckAllowSecureChannelModal from "./SoftwareCheckAllowSecureChannelModal";
 import GenuineCheckCancelModal from "./GenuineCheckCancelModal";
 import GenuineCheckNotGenuineModal from "./GenuineCheckNotGenuineModal";
 import { Status as SoftwareCheckStatus } from "./shared";
@@ -20,6 +22,7 @@ export type Props = {
   isDisplayed?: boolean;
   onComplete: () => void;
   productName: string;
+  deviceModelId: DeviceModelId;
 };
 
 // The commands needs to be defined outside of the function component, to avoid creating it at
@@ -29,7 +32,7 @@ const getLatestAvailableFirmwareFromDeviceIdCommand = command(
   "getLatestAvailableFirmwareFromDeviceId",
 );
 
-const SoftwareCheckStep = ({ isDisplayed, onComplete, productName }: Props) => {
+const SoftwareCheckStep = ({ isDisplayed, onComplete, productName, deviceModelId }: Props) => {
   const [genuineCheckStatus, setGenuineCheckStatus] = useState<SoftwareCheckStatus>(
     SoftwareCheckStatus.inactive,
   );
@@ -41,13 +44,13 @@ const SoftwareCheckStep = ({ isDisplayed, onComplete, productName }: Props) => {
   const device = useSelector(getCurrentDevice);
   const deviceId = device?.deviceId ?? "";
 
-  const { genuineState, devicePermissionState, error, resetGenuineCheckState } = useGenuineCheck({
+  const { genuineState, devicePermissionState, resetGenuineCheckState } = useGenuineCheck({
     getGenuineCheckFromDeviceId: getGenuineCheckFromDeviceIdCommand,
     isHookEnabled: genuineCheckStatus === SoftwareCheckStatus.active,
     deviceId,
   });
 
-  const { latestFirmware /*, error */, status } = useGetLatestAvailableFirmware({
+  const { latestFirmware, status, lockedDevice } = useGetLatestAvailableFirmware({
     getLatestAvailableFirmwareFromDeviceId: getLatestAvailableFirmwareFromDeviceIdCommand,
     isHookEnabled: firmwareUpdateStatus === SoftwareCheckStatus.active,
     deviceId,
@@ -108,6 +111,26 @@ const SoftwareCheckStep = ({ isDisplayed, onComplete, productName }: Props) => {
     }
   }, [isDisplayed, firmwareUpdateStatus, onComplete, status, genuineCheckStatus, latestFirmware]);
 
+  const lockedDeviceOnClose = useCallback(() => {
+    if (genuineCheckStatus === SoftwareCheckStatus.active) {
+      // Triggers a prompt asking for the user to retry or confirm cancelling
+      setGenuineCheckStatus(SoftwareCheckStatus.cancelled);
+    } else if (firmwareUpdateStatus === SoftwareCheckStatus.active) {
+      // No prompt, directly failed status
+      setFirmwareUpdateStatus(SoftwareCheckStatus.failed);
+    }
+  }, [firmwareUpdateStatus, genuineCheckStatus]);
+
+  const lockedDeviceModalIsOpen =
+    (devicePermissionState === "unlock-needed" &&
+      genuineCheckStatus === SoftwareCheckStatus.active) ||
+    (lockedDevice && firmwareUpdateStatus === SoftwareCheckStatus.active);
+
+  const allowSecureChannelIsOpen =
+    devicePermissionState === "requested" &&
+    (genuineCheckStatus === SoftwareCheckStatus.active ||
+      firmwareUpdateStatus === SoftwareCheckStatus.active);
+
   return (
     <Box>
       <GenuineCheckModal
@@ -138,12 +161,15 @@ const SoftwareCheckStep = ({ isDisplayed, onComplete, productName }: Props) => {
           setGenuineCheckStatus(SoftwareCheckStatus.failed);
         }}
       />
-      <GenuineCheckAnimationModal
-        isOpen={devicePermissionState === "unlock-needed" || devicePermissionState === "requested"}
-        animationName={
-          devicePermissionState === "unlock-needed" ? "plugAndPinCode" : "allowManager"
-        }
-        deviceId={device?.modelId}
+      <SoftwareCheckLockedDeviceModal
+        isOpen={lockedDeviceModalIsOpen}
+        deviceModelId={deviceModelId}
+        productName={productName}
+        onClose={lockedDeviceOnClose}
+      />
+      <SoftwareCheckAllowSecureChannelModal
+        isOpen={allowSecureChannelIsOpen}
+        deviceModelId={deviceModelId}
         productName={productName}
       />
       <SoftwareCheckContent