+++ apps/ledger-live-mobile/src/actions/settings.ts
@@ -2,7 +2,12 @@ import { useCallback } from "react";
 import { useTranslation } from "react-i18next";
 import { createAction } from "redux-actions";
 import { useDispatch, useSelector } from "react-redux";
-import { DeviceModelInfo, DeviceInfo } from "@ledgerhq/types-live";
+import {
+  DeviceModelInfo,
+  DeviceInfo,
+  FeatureId,
+  Feature,
+} from "@ledgerhq/types-live";
 import type { Device } from "@ledgerhq/live-common/hw/actions/types";
 import type { PortfolioRange } from "@ledgerhq/types-live";
 import { MarketListRequestParams } from "@ledgerhq/live-common/market/types";
@@ -19,7 +24,7 @@ import {
   SettingsAcceptSwapProviderPayload,
   SettingsAddStarredMarketcoinsPayload,
   SettingsBlacklistTokenPayload,
-  SettingsDangerouslyOverrideStatePayload,
+  DangerouslyOverrideStatePayload,
   SettingsDismissBannerPayload,
   SettingsHideEmptyTokenAccountsPayload,
   SettingsHideNftCollectionPayload,
@@ -63,6 +68,10 @@ import {
   SettingsActionTypes,
   SettingsSetWalletTabNavigatorLastVisitedTabPayload,
   SettingsSetDismissedDynamicCardsPayload,
+  SettingsSetStatusCenterPayload,
+  SettingsSetOverriddenFeatureFlagPlayload,
+  SettingsSetOverriddenFeatureFlagsPlayload,
+  SettingsSetFeatureFlagsBannerVisiblePayload,
 } from "./types";
 import { WalletTabNavigatorStackParamList } from "../components/RootNavigator/types/WalletTabNavigator";
 
@@ -485,8 +494,45 @@ export const setWalletTabNavigatorLastVisitedTab = (
     walletTabNavigatorLastVisitedTab,
   });
 
+const setStatusCenterAction = createAction<SettingsSetStatusCenterPayload>(
+  SettingsActionTypes.SET_STATUS_CENTER,
+);
+export const setStatusCenter = (displayStatusCenter: boolean) =>
+  setStatusCenterAction({
+    displayStatusCenter,
+  });
+
+const setOverriddenFeatureFlagAction =
+  createAction<SettingsSetOverriddenFeatureFlagPlayload>(
+    SettingsActionTypes.SET_OVERRIDDEN_FEATURE_FLAG,
+  );
+export const setOverriddenFeatureFlag = (
+  id: FeatureId,
+  value: Feature | undefined,
+) =>
+  setOverriddenFeatureFlagAction({
+    id,
+    value,
+  });
+
+const setOverriddenFeatureFlagsAction =
+  createAction<SettingsSetOverriddenFeatureFlagsPlayload>(
+    SettingsActionTypes.SET_OVERRIDDEN_FEATURE_FLAGS,
+  );
+export const setOverriddenFeatureFlags = (overriddenFeatureFlags: {
+  [key in FeatureId]?: Feature;
+}) => setOverriddenFeatureFlagsAction({ overriddenFeatureFlags });
+
+const setFeatureFlagsBannerVisibleAction =
+  createAction<SettingsSetFeatureFlagsBannerVisiblePayload>(
+    SettingsActionTypes.SET_FEATURE_FLAGS_BANNER_VISIBLE,
+  );
+export const setFeatureFlagsBannerVisible = (
+  featureFlagsBannerVisible: boolean,
+) => setFeatureFlagsBannerVisibleAction({ featureFlagsBannerVisible });
+
 const dangerouslyOverrideStateAction =
-  createAction<SettingsDangerouslyOverrideStatePayload>(
+  createAction<DangerouslyOverrideStatePayload>(
     SettingsActionTypes.DANGEROUSLY_OVERRIDE_STATE,
   );
 export const dangerouslyOverrideState = (s: State) =>