+++ apps/ledger-live-mobile/src/components/RootNavigator/types/ClaimNftNavigator.tsx
@@ -1,0 +1,6 @@
+import { ScreenName } from "../../../const";
+
+export type ClaimNftNavigatorParamList = {
+  [ScreenName.ClaimNftWelcome]: undefined;
+  [ScreenName.ClaimNftQrScan]: undefined;
+};