+++ apps/ledger-live-mobile/src/components/RootNavigator/types/RequestAccountNavigator.ts
@@ -1,13 +1,11 @@
 import { WalletAPIAccount } from "@ledgerhq/live-common/wallet-api/types";
 import type {
-  CryptoCurrency,
   CryptoOrTokenCurrency,
-  Currency,
   TokenCurrency,
 } from "@ledgerhq/types-cryptoassets";
 
 import type { Account, AccountLike } from "@ledgerhq/types-live";
-import { Observable } from "@ledgerhq/wallet-api-server";
+import { Observable } from "rxjs7";
 import { NavigatorScreenParams } from "@react-navigation/native";
 import { NavigatorName, ScreenName } from "../../../const";
 import { AddAccountsNavigatorParamList } from "./AddAccountsNavigator";
@@ -15,14 +13,14 @@ import { AddAccountsNavigatorParamList } from "./AddAccountsNavigator";
 export type RequestAccountNavigatorParamList = {
   [ScreenName.RequestAccountsSelectCrypto]: {
     accounts$?: Observable<WalletAPIAccount[]>;
-    currencies: Currency[];
+    currencies: CryptoOrTokenCurrency[];
     allowAddAccount?: boolean;
     onSuccess?: (account: AccountLike, parentAccount?: Account) => void;
     onError?: (_: Error) => void;
   };
   [ScreenName.RequestAccountsSelectAccount]: {
     accounts$?: Observable<WalletAPIAccount[]>;
-    currency: CryptoCurrency | TokenCurrency;
+    currency: CryptoOrTokenCurrency;
     allowAddAccount?: boolean;
     onSuccess?: (account: AccountLike, parentAccount?: Account) => void;
     onError?: (_: Error) => void;