+++ apps/ledger-live-mobile/src/components/WebViewScreen/index.tsx
@@ -35,6 +35,7 @@ export type Props = {
   renderHeader?: () => ReactNode;
   renderLoading?: () => ReactNode;
   renderError?: () => JSX.Element;
+  enableNavigationOverride?: boolean;
 };
 
 const WebViewScreen = ({
@@ -45,6 +46,7 @@ const WebViewScreen = ({
   renderHeader,
   renderLoading,
   renderError,
+  enableNavigationOverride = true,
 }: Props) => {
   const ref = useRef<WebView>(null);
   const navigation = useNavigation();
@@ -63,6 +65,7 @@ const WebViewScreen = ({
   }, [uri, setLoading]);
 
   useEffect(() => {
+    if (!enableNavigationOverride) return;
     const unsubscribe = navigation.addListener("beforeRemove", e => {
       if (canGoBack) return;
       // Prevent default behavior of leaving the screen
@@ -70,6 +73,7 @@ const WebViewScreen = ({
       ref.current?.goBack();
     });
 
+    // eslint-disable-next-line consistent-return
     return unsubscribe;
   }, [canGoBack, navigation]);