+++ apps/ledger-live-mobile/src/const/navigation.ts
@@ -15,6 +15,8 @@ export enum ScreenName {
   Assets = "Assets",
   BenchmarkQRStream = "BenchmarkQRStream",
   BitcoinEditCustomFees = "BitcoinEditCustomFees",
+  ClaimNftWelcome = "ClaimNftWelcome",
+  ClaimNftQrScan = "ClaimNftQrScan",
   ClaimRewardsConnectDevice = "ClaimRewardsConnectDevice",
   ClaimRewardsSelectDevice = "ClaimRewardsSelectDevice",
   ClaimRewardsValidationError = "ClaimRewardsValidationError",
@@ -42,6 +44,7 @@ export enum ScreenName {
   DebugInformation = "DebugInformation",
   DebugLogs = "DebugLogs",
   DebugLottie = "DebugLottie",
+  DebugTermsOfUse = "DebugTermsOfUse",
   DebugMockGenerateAccounts = "DebugMockGenerateAccounts",
   DebugNetwork = "DebugNetwork",
   DebugSettings = "DebugSettings",
@@ -408,7 +411,7 @@ export enum ScreenName {
   WalletConnectScan = "WalletConnectScan",
   WalletConnectConnect = "WalletConnectConnect",
   WalletConnectDeeplinkingSelectAccount = "WalletConnectDeeplinkingSelectAccount",
-  NotificationCenterNews = "NotificationCenterNews",
+  NotificationCenter = "NotificationCenter",
   NotificationCenterStatus = "NotificationCenterStatus",
   SignTransactionSummary = "SignTransactionSummary",
   SignTransactionSelectDevice = "SignTransactionSelectDevice",
@@ -431,6 +434,7 @@ export enum ScreenName {
   NoDeviceWallScreen = "NoDeviceWallScreen",
   DiscoverScreen = "DiscoverScreen",
   Learn = "Learn",
+  LearnWebView = "LearnWebView",
   // cardano
   CardanoEditMemo = "CardanoEditMemo",
   // hedera
@@ -463,6 +467,9 @@ export enum ScreenName {
   CustomImagePreviewPreEdit = "CustomImagePreviewPreEdit",
   CustomImagePreviewPostEdit = "CustomImagePreviewPostEdit",
   CustomImageNFTGallery = "CustomImageNFTGallery",
+
+  ClaimNftWelcomeScreen = "ClaimNftWelcomeScreen",
+  ClaimNftQrScanScreen = "ClaimNftQrScanScreen",
 
   PostOnboardingHub = "PostOnboardingHub",
   PostOnboardingDebugScreen = "PostOnboardingDebugScreen",
@@ -560,6 +567,9 @@ export enum NavigatorName {
 
   // Custom Image
   CustomImage = "CustomImage",
+
+  // Claim NFT
+  ClaimNft = "ClaimNft",
 
   WalletConnect = "WalletConnect",