+++ apps/ledger-live-mobile/src/screens/PostOnboarding/PostOnboardingHub.tsx
@@ -16,6 +16,7 @@ import {
   useAllPostOnboardingActionsCompleted,
   usePostOnboardingHubState,
 } from "@ledgerhq/live-common/postOnboarding/hooks/index";
+import { PostOnboardingActionId } from "@ledgerhq/types-live";
 import { clearPostOnboardingLastActionCompleted } from "@ledgerhq/live-common/postOnboarding/actions";
 import { useDispatch } from "react-redux";
 import { getDeviceModel } from "@ledgerhq/devices";
@@ -28,6 +29,7 @@ import {
 } from "../../components/RootNavigator/types/helpers";
 import { PostOnboardingNavigatorParamList } from "../../components/RootNavigator/types/PostOnboardingNavigator";
 import DeviceSetupView from "../../components/DeviceSetupView";
+import { useCompleteActionCallback } from "../../logic/postOnboarding/useCompleteAction";
 
 const AnimatedFlex = Animated.createAnimatedComponent(Flex);
 
@@ -38,10 +40,11 @@ type NavigationProps = BaseComposite<
   >
 >;
 
-const PostOnboardingHub = ({ navigation }: NavigationProps) => {
+const PostOnboardingHub = ({ navigation, route }: NavigationProps) => {
   const dispatch = useDispatch();
   const { t } = useTranslation();
   const { actionsState, deviceModelId } = usePostOnboardingHubState();
+  const completePostOnboardingAction = useCompleteActionCallback();
 
   const clearLastActionCompleted = useCallback(() => {
     dispatch(clearPostOnboardingLastActionCompleted());
@@ -55,6 +58,20 @@ const PostOnboardingHub = ({ navigation }: NavigationProps) => {
      * */
     () => clearLastActionCompleted,
     [clearLastActionCompleted],
+  );
+
+  useEffect(
+    /**
+     * Complete claim NFT action if the route param completed is true
+     * */
+    () => {
+      route &&
+        route.params &&
+        route.params.completed &&
+        route.params.completed === "true" &&
+        completePostOnboardingAction(PostOnboardingActionId.claimNft);
+    },
+    [clearLastActionCompleted, completePostOnboardingAction, route],
   );
 
   const allowClosingScreen = useRef<boolean>(true);
@@ -77,8 +94,10 @@ const PostOnboardingHub = ({ navigation }: NavigationProps) => {
 
   const animationTimeout = useRef<ReturnType<typeof setTimeout> | null>(null);
   const clearAnimationTimeout = useCallback(() => {
-    animationTimeout.current && clearTimeout(animationTimeout.current);
-  }, [animationTimeout]);
+    !allDone &&
+      animationTimeout.current &&
+      clearTimeout(animationTimeout.current);
+  }, [allDone]);
 
   const triggerEndAnimation = useCallback(() => {
     const onAnimEnd = () => {