+++ libs/ledger-live-common/src/wallet-api/helpers.ts
@@ -6,11 +6,27 @@ import {
   WalletAPIERC20TokenCurrency,
   WalletAPICryptoCurrency,
 } from "./types";
+import { WALLET_API_FAMILIES } from "./constants";
+
+// Small helper to avoid issues with includes and typescript
+// more infos: https://fettblog.eu/typescript-array-includes/
+function includes<T extends U, U>(
+  array: ReadonlyArray<T>,
+  element: U
+): element is T {
+  return array.includes(element as T);
+}
 
 export function isWalletAPISupportedCurrency(
   currency: Currency
 ): currency is WalletAPISupportedCurrency {
-  return isCryptoCurrency(currency) || isTokenCurrency(currency);
+  if (isCryptoCurrency(currency)) {
+    return includes(WALLET_API_FAMILIES, currency.family);
+  }
+  if (isTokenCurrency(currency)) {
+    return includes(WALLET_API_FAMILIES, currency.parentCurrency.family);
+  }
+  return false;
 }
 
 export function isWalletAPICryptoCurrency(