+++ libs/ledgerjs/packages/devices/src/helpers.ts
@@ -1,0 +1,14 @@
+import { DeviceModelId } from ".";
+
+export const stringToDeviceModelId = (
+  strDeviceModelId: string,
+  defaultDeviceModelId: DeviceModelId
+): DeviceModelId => {
+  if (
+    Object.values(DeviceModelId)?.includes(strDeviceModelId as DeviceModelId)
+  ) {
+    return DeviceModelId[strDeviceModelId as DeviceModelId];
+  }
+
+  return defaultDeviceModelId;
+};