+++ libs/ledgerjs/packages/types-live/src/feature.ts
@@ -2,6 +2,7 @@
 export type FeatureId =
   | "learn"
   | "brazePushNotifications"
+  | "brazeLearn"
   | "llmNewDeviceSelection"
   | "llmUsbFirmwareUpdate"
   | "ratingsPrompt"
@@ -41,7 +42,8 @@ export type FeatureId =
   | "postOnboardingClaimNft"
   | "postOnboardingAssetsTransfer"
   | "firebaseEnvironmentReadOnly"
-  | "protectServicesMobile";
+  | "protectServicesMobile"
+  | "swapShowDexQuotes";
 
 /**  We use objects instead of direct booleans for potential future improvements
 like feature versioning etc */